refactor: moved CPU's _alu8 field to Instructions8
